Add dest parameter.
Use when deploying a project from a remote host. Requires docker_service plugin. When project_src and
dest are provided, contents of project_src will be recursively copied from the control host to dest on
the target host. The module is then executed with project_src set to the value of dest.

Only contents of project_src are copied. If the docker-compose.yml file (other other project files)
reference files outside of project_src, those files will not be copied.

Expects dest to be the final directory name. For example, if project_src basename is *foo*, and the
desired basename of dest is *foo*, include *foo* in dest value.